requirements.txt
beautifulsoup4==4.11.1
fastapi==0.115.12
geopy==2.4.1
passlib==1.7.4
pycountry==22.3.5
pydantic==2.11.5
pydantic_mongo==3.1.0
PyJWT==2.10.1
pymongo==4.13.0
python-decouple==3.8
Requests==2.32.4
SocksiPy_branch==1.01
urllib3==2.5.0
uvicorn==0.34.2
PySocks==1.7.1
email_validator==2.2.0
python-multipart==0.0.20

# Testing dependencies
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
mongomock==4.1.2
httpx==0.26.0